% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{convenience}
\alias{convenience}
\alias{important_}
\alias{url_}
\alias{rgb_}
\alias{rgba_}
\title{Convenience Functions}
\usage{
important_(value)

url_(value)

rgb_(r, g, b)

rgba_(r, g, b, a = 0.5)
}
\arguments{
\item{value}{Value to use.}

\item{r, g, b, a}{Red, green, blue, and alpha values.}
}
\description{
Convenience functions for common operations.
}
\examples{
Linne$
 new()$
 change(
   sel_id("id"),
   color = rgba(255, 255, 255, .6)
 )

}
