% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{pipes}
\alias{pipes}
\alias{\%child\%}
\alias{\%or\%}
\alias{\%and\%}
\title{Infixes}
\usage{
lhs \%child\% rhs

lhs \%or\% rhs

lhs \%and\% rhs
}
\arguments{
\item{lhs, rhs}{Selectors as returned by \verb{sel_*} family of functions.}
}
\description{
Convenient pipes for more sophisticated selectors.
}
\section{Operators}{

\itemize{
\item \code{\link{\%child\%}} - Adds left hand as child of right hand.
\item \code{\link{\%or\%}} - Select left hand or right hand, ideal to select multiple elements at once.
\item \code{\link{\%and\%}} - Select left hand and right hand, ideal to narrow down the selection.
}
}

\examples{
# select all paragraph 'p' where class = "red"
sel_tag("p") \%and\% sel_class("red")

# the other way around works equally well
sel_class("red") \%and\% sel_tag("p")

# select multiple elements
# where id = "x" or class = "center" 
sel_id("x") \%or\% sel_class("center")

# select element with id = "x" and parent's id = "y"
sel_id("y") \%child\% sel_id("y")

}
