% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/when.R
\name{when}
\alias{when}
\alias{when_active}
\alias{when_hover}
\alias{when_focus}
\title{State}
\usage{
when_active(selector)

when_hover(selector)

when_focus(selector)
}
\arguments{
\item{selector}{as returned by \link{selectors}.}
}
\description{
Narrows selection to a specific state, e.g.: when it is hovered.
}
