% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{convenience}
\alias{convenience}
\alias{important_}
\alias{url_}
\alias{rgb_}
\alias{rgba_}
\title{Convenience Functions}
\usage{
important_(value)

url_(value)

rgb_(r, g, b)

rgba_(r, g, b, a = 0.5)
}
\arguments{
\item{value}{Value to use.}

\item{r, g, b, a}{Red, green, blue, and alpha values.}
}
\description{
Convenience functions for common operations.
}
\section{Functions}{

\itemize{
\item \code{\link[=important_]{important_()}} - Makes it such that the rule cannot be overwritten by other rules (other selections).
\item \code{\link[=rgb_]{rgb_()}}, \code{\link[=rgba_]{rgba_()}} - Functions for red, green, blue and alpha for transparency.
\item \code{\link[=url_]{url_()}} - Wrap in a \code{url} CSS function call.
}
}

\examples{
Linne$
 new()$
 rule(
   sel_id("id"),
   color = rgba_(255, 255, 255, .6),
   fontSize = important_(20)
 )

}
